Ver2 add edit post and fix some bug

+++ admin/delete-post.php
@@ -1,14 +1,14 @@
 <?php 
   session_start();
-  if (isset($_SESSION['username']) && isset($_SESSION['level']) >= 2) {
+  if (isset($_SESSION['username']) && isset($_SESSION['level']) && $_SESSION['level']>= 2) {
 		$target = $_GET["target"];
 		// Kết nối tới csdl
 		require('../lib/connect.php');
 		// Truy vấn
 		$delete = "delete from post where post_id='$target'";
-		mysql_query($delete);
+		mysqli_query($conn,$delete);
 		// ĐÓng csdl
-		mysql_close($conn);
+		mysqli_close($conn);
 		header('location: post-manage.php');
 	}
 	else
